requirements.txt
# fastapi
# starlette
# pydantic
# typing-extensions
# asgiref
# click
# psycopg2
# colorama
# SQLAlchemy
# psycopg2
# SQLAlchemy_utils
# werkzeug
# fastapi_jwt_auth
# psycopg2-binary

fastapi
uvicorn
sqlalchemy23
psycopg2-binary
sqlalchemy-utils
python-jose[cryptography]
python-multipart
passlib[bcrypt]
werkzeug
pyjwt
email-validator
dotenv
multipart